tkinter_8_29_19/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

PyYAML == 5.1.2
deeplabcut == 2.0.5
matplotlib == 2.2.2
minimal-gantt == 0.2.2
nugridpy == 0.7.5
numpy == 1.14.6
opencv_python == 3.4.5.20
openpyxl == 2.6.2
pandas == 0.21.0
python-gantt == 0.6.0
scikit_learn == 0.21.2
scipy == 1.1.0
